Conflicts:
	src/columns/Publisher.cpp
	src/connections/TransposePoolingConn.cpp
	tests/AvgPoolTest/src/AvgPoolTestInputLayer.hpp
	tests/DelaysToFeaturesTest/src/DelayTestProbe.cpp
	tests/GenericSystemTest/src/main.cpp
	tests/ImprintConnTest/src/InputLayer.hpp
	tests/InputSystemTest/.project
	tests/InputSystemTest/linux-cmake.cproject
	tests/InputSystemTest/linux-eclipse.cproject
	tests/InputSystemTest/mac-cmake.cproject
	tests/LCATest/src/main.cpp
	tests/MaskLayerTest/src/InputLayer.hpp
	tests/MomentumLCATest/src/main.cpp
	tests/SumPoolTest/src/SumPoolTestInputLayer.hpp
	tests/TotalEnergyTest/src/main.cpp
	tests/WriteActivitySparseTest/src/WriteActivitySparseTest.cpp
	tests/WriteSparseFileTest/src/main.cpp
	tests/test_border_activity/src/test_border_activity.cpp

+++ src/layers/accumulate_functions.c
@@ -98,16 +98,13 @@ int pvpatch_max_pooling(int kPreGlobalExt, int nk, float* RESTRICT v, float a, p
 {
   int k;
   int err = 0;
-  //float * gate = (float *)auxPtr;
-  int * gate = (int*)auxPtr;
+  float * gate = (float *)auxPtr;
   for (k = 0; k < nk; k+=sf) {
-    //     v[k] = v[k] > a*w[k] ? v[k] : a*w[k];
-    //v[k] = v[k] > a*w[0] ? v[k] : a*w[0];
     float checkVal = a*w[0];
     if(v[k] <= checkVal){
        v[k] = checkVal;
        if(gate){
-          gate[k] = kPreGlobalExt;
+          gate[k] = (float) kPreGlobalExt;
        }
     }
   }
@@ -118,14 +115,12 @@ int pvpatch_max_pooling_from_post(int kPreGlobalExt, int nk, float * RESTRICT v,
    int status = 0;
    int k;
    float vmax = *v;
-   int* gate = (int*)auxPtr;
+   float* gate = (float*)auxPtr;
    int gateMax;
    if(gate){
-      gateMax = *gate;
+      gateMax = (int) *gate;
    }
    for (k = 0; k < nk; k+=sf) {
-      //vmax = vmax > a[k]*w[k] ? vmax : a[k]*w[k];
-      //vmax = vmax > a[k]*w[0] ? vmax : a[k]*w[0];
       float checkVal = a[k]*w[0];
       if(vmax <= checkVal){
          vmax = checkVal;
@@ -136,7 +131,7 @@ int pvpatch_max_pooling_from_post(int kPreGlobalExt, int nk, float * RESTRICT v,
    }
    *v = vmax;
    if(gate){
-      *gate = gateMax;
+      *gate = (float) gateMax;
    }
    return status;
 }
